requirements.txt
hikari-yuyo~=1.19.0
aiobungie[speedup]~=0.2.10
hikari[speedups]~=2.0.0.dev122
hikari-tanjun~=2.17.1
hikari-yuyo~=1.19.0a1
asyncpg~=0.29.0
redis~=5.0.1
click~=8.1.7
